refactor(renderer): move app shell core out of elements/ into app/
Step 4 of the renderer reorganization. The Moss runtime shell — its
top-level component, navigation chrome and app-level dialogs — moves into
app/:
- app/main-dashboard.ts
- app/navigation/  — groups-sidebar, sidebar-button, group-sidebar-button,
  group-applets-row (joining applet-sidebar-button + group-area-sidebar)
- app/dialogs/     — create-group, join-group, select-group, loading,
  tool-info dialogs

Pure file moves + import-path updates; no behaviour change.
