+++ src/renderer/src/app/dialogs/join-group-dialog.ts
