+++ src/renderer/src/app/dialogs/loading-dialog.ts
