+++ src/renderer/src/app/dialogs/select-group-dialog.ts
