+++ src/renderer/src/app/navigation/group-applets-row.ts
