+++ src/renderer/src/app/navigation/group-sidebar-button.ts
