+++ src/renderer/src/app/navigation/groups-sidebar.ts
