+++ src/renderer/src/app/navigation/sidebar-button.ts
