+++ src/renderer/src/elements/creatables/creatable-panel.ts
@@ -25,7 +25,7 @@ import { MossStore } from '../../moss-store.js';
 import '../pocket/wal-element.js';
 import '../pocket/pocket-search.js';
 import './creatable-view.js';
-import '../navigation/group-applets-row.js';
+import '../../app/navigation/group-applets-row.js';
 
 import { StoreSubscriber } from '@holochain-open-dev/stores';
