+++ src/renderer/src/elements/debugging-panel/app-debugging-details.ts
@@ -7,7 +7,7 @@ import '@shoelace-style/shoelace/dist/components/card/card.js';
 
 import '../../groups/elements/group-context.js';
 import '../../applets/elements/applet-logo.js';
-import '../dialogs/create-group-dialog.js';
+import '../../app/dialogs/create-group-dialog.js';
 import '../reusable/groups-for-applet.js';
 import './cell-details.js';
