+++ src/renderer/src/elements/debugging-panel/state-dump.ts
@@ -22,7 +22,7 @@ import '@shoelace-style/shoelace/dist/components/button/button.js';
 
 import '../../groups/elements/group-context.js';
 import '../../applets/elements/applet-logo.js';
-import '../dialogs/create-group-dialog.js';
+import '../../app/dialogs/create-group-dialog.js';
 import '../reusable/groups-for-applet.js';
 
 import { mossStyles } from '../../shared-styles.js';
