+++ src/renderer/src/personal-views/tool-library/elements/installable-tools.ts
@@ -17,7 +17,7 @@
 // import { mossStyles } from '../../../shared-styles.js';
 // import { MossStore } from '../../../moss-store.js';
 // import { mossStoreContext } from '../../../context.js';
-// import { SelectGroupDialog } from '../../../elements/dialogs/select-group-dialog.js';
+// import { SelectGroupDialog } from '../../../app/dialogs/select-group-dialog.js';
 // import '../../../elements/dialogs/select-group-dialog.js';
 // import TimeAgo from 'javascript-time-ago';
 // import './tool-publisher.js';
